asus-nb-wmi: Map 0x35 to KEY_SCREENLOCK
On Windows, with the ATK ACPI driver from Asus installed, the display
toggle hotkey (Fn+F7) turns the laptop screen off, with a fading effect.
If external displays are connected, those are turned off too. Pressing a
key or moving the cursor wakes-up the display, with the key press or
cursor click passed through to the active application (i.e. in addition
to turning the displays back on, the input event is handled as normal).

Mapping it to KEY_SCREENLOCK can be used to trigger the window manager
to lock the screen, achieving a similar behavior.

Signed-off-by: João Paulo Rechi Vita <jprvita@endlessm.com>

https://phabricator.endlessm.com/T20914